requirements.txt
asgiref==3.4.1
Django==3.2.5
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
